fix(donnees-personnelles): type de retour de la suppression derivee des notifications
deleteByDestinataireId rendait un long. Une methode annotee @Modifying n'accepte que void
ou int / Integer (ModifyingExecution le verifie a la creation du depot) : selon le chemin
d'execution choisi par Spring Data pour une requete derivee de suppression, le demarrage sous
le profil postgres pouvait echouer. Une suppression Spring Data rend de toute facon un int.

+++ src/main/java/dz/tabibi/backend/notifications/adapter/NotificationJpa.java
@@ -12,7 +12,11 @@ interface NotificationJpa extends JpaRepository<NotificationEntity, UUID> {
 
     long countByDestinataireIdAndLueFalse(UUID destinataireId);
 
-    /** Effacement de compte : requete derivee de suppression, a executer dans une transaction. */
+    /**
+     * Effacement de compte : requete derivee de suppression, a executer dans une transaction.
+     * Le type de retour est {@code int} : c'est ce que rend une suppression Spring Data, et c'est
+     * le seul type qu'accepte l'execution d'une requete annotee {@link Modifying}.
+     */
     @Modifying
-    long deleteByDestinataireId(UUID destinataireId);
+    int deleteByDestinataireId(UUID destinataireId);
 }
